requirements.txt
python-telegram-bot
flask
pandas
fpdf2
gunicorn
Flask-SQLAlchemy
psycopg2-binary